Sidebar with info.
Click on the mark. Sidebar will popup from right side and hide the
layer switch on the right side. Click on the map or the x close button
to close the side bar.
Configuration of the Content of the sidebar can be done in different
ways. Feel free to give suggestions. Cheers…
Cleber